onboarding: remove check for account discoverable in suggestions. (#134)
As we've secured permission for all the suggested follows currently in the YAML, we should be able to present them, even though some accounts have 'discoverable' set to false on Mastodon. We may consider a system in the future which checks the field on a case-by-case basis for accounts where it's required.